Add files via upload
This produces plots showing the average p-value and average p-value + two standard deviations for both cosmic times and mass ratios of massive black hole mergers, as a function of the number of massive black hole mergers observed by LISA over the course of the mission. The significance threshold (p-value of 0.05) is denoted by a red line.