Add files via upload
This script computes the average p-value and the standard deviation for the p-value when comparing the mass ratios of massive black hole mergers with dark matter that has non-gravitational self-interactions to dark matter that does not. The KS test is used to compute the p-values. A line near the top can be edited to change the number of mergers assumed to be observed by LISA over the course of the mission. This is a parameter we vary. Note that this file and the cosmic times pipeline both contain the data from the simulations, namely the cosmic times and mass ratios of massive black hole mergers in the simulated galaxies.