Add files via upload
This script plots SNR (signal-to-noise) ranges as seen by LISA for the mergers in Galaxy GM7 with CDM (dark matter without non-gravitational self-interactions) and SIDM (dark matter with non-gravitational self-interactions, with a constant cross section per unit mass of 1 sq cm per gram). This script varies inclination and spins over the full range of possible values, as these values are not provided by the simulations.